refactor fonctions
