--- src/app/[locale]/lib/parse-url.tsx
@@ -1,12 +1,0 @@
-export const parseURL = (url: string): string => {
-	const parsedURL = URL.parse(url);
-	if (parsedURL) {
-		if (parsedURL.hostname.startsWith("www.")) {
-			return parsedURL.hostname.slice(4);
-		}
-
-		return parsedURL.hostname;
-	}
-
-	return url;
-};
