--- src/app/[locale]/lib/sanitize-html.ts
@@ -1,8 +1,0 @@
-import DOMPurify from "isomorphic-dompurify";
-
-export function sanitizeHTML(html: string): string {
-	return DOMPurify.sanitize(html, {
-		ALLOWED_TAGS: ["p", "a", "b", "i", "em", "strong", "br"],
-		ALLOWED_ATTR: ["href"],
-	});
-}
